requirements.txt
django==2.2
djangorestframework==3.12.1  
django-cors-headers==3.5.0

